Corresponding arrays' sum
Finds the bigger of the two arrays, then creates another array based on the largest array. Then adds the elements in corresponding indices of the first and second array and stores the result respectively in the third dynamically created array.
Constraints followed:
If the size is 0 or less print "Invalid array size".