Array Compatibility Check
Two arrays are said to be compatible if they are of the same size and if the ith element in the first array is greater than or equal to the ith element in the second array for all i elements. 
Constraints followed:
If the array size is zero or lesser then display the message "Invalid array size".