Grade caculator
Creates a class Student with following private attribute :
int id, String name, marks(integer array), float average and char grade. Getter and setter methods included for all the attributes.
void calculateAvg() method calculates average and set average mark for the current student.
void findGrade() method sets the grade based on the average calculated.
All marks should be greater than or equal to 50 to calculate the grade. If one or more marks is below 50 the grade is set as 'F'. If all marks are 50 and above, the grade is based on the below conditions, 
If the average is between 80 and 100 then, then return grade as 'O'. If the average is between 50 and 79 then return the grade as  'A' .If the average of the student is  less than 50 then return grade as 'F'.
(Note : Assume the number of subject is  greater than zero and  the mark for a subject is in the range of 0 - 100.)
(StudentMain class is the driver.)