PAN card validation
Validates PAN cards according to their standard format. Which is as follows:
PAN Number should have exactly ten characters.  First 5 characters should be alphabets in upper case, next four characters should be digits from 0 to 9 and the last character should be an alphabet in capital case.
Uses regex instead of tediously iterating through the character array and finding a match with a duplicate array converted to uppercase, and then based on index, finding if it matches the criteria, even though that can be done and would work as accurately as regex.