Coffee Bar Numerology
A string is considered as numerological if the entered string should be strictly alphabetical and the number of characters in it should be an even number. (According to a question I read. I still have yet to know what they do in numerology myself). This program checks if the given string is numerological or not.