Sorts 0 1 2 elements without any sorting algorithm
Method: checks for 0 then 1 else 2 and swaps. Has low for the 0 elements, mid for the 1 elements and 2 for the high elements. 
This is almost the same as the solution given in GFG. 

The initial solution I did was a simple if condition then append to a new array which would give the elements in the initial array, but sorted. The question was to do that in-place. Hence the above solution works better. In any case, the code for the "if condition new array append" is also given below the actual solution.