Cyclically rotates the array by 1
Takeaways:
Tried to think of a way to replace the i+1th element with the ith element but that would cause a problem in the next iteration of the for loop as the i+1th current element would actually be the ith element, thereby it would just replace all the elements of the array (or I should say list since this is python) with the same element. Thought of a way to store the i+1 element itself as a temp value and paste it for the next iteration's i+1 (essentially i+2), but that would cause problems of its own.

The most simple approach is often the best approach. Reverse traversal.
Stored the last element since that's always the element that gets put at the beginning of the cycled array, and then just replaced the a[i(from the last)] with the 
a[i-1]th element. When the iteration eventually reaches a[0], we just replace it with the last element we had stored from earlier and BOOM, we have a Cyclically rotated array :)

Cool problem!