Minimum jumps required to traverse
While traversing, lets say the first element is 2, so you jump 2 positions and if the element you find there after jumping is 5, you jump 5 positions. Like this, how many jumps are required to traverse the array is this problem. 

Logic:
When we arrive at a position, we check the value and give it to the iteration value i. We give the value to iteration only if i + value <= length of array. Otherwise we treat it as a single jump and break off from the loop. We record each jump with a separate variable (called jump here).