Find a duplicate and it's location
This is to find a duplicate of a number (assuming only 1 number has a duplicate in the array) and where that actual duplicate is in the array.

Logic:
Use a searching method on the array and do that for each element in the array. Take one element, search the rest of the array. Take the next element and search the rest of the array and so on.
 
Will improve upon this code further later by adding functionality to search for multiple duplicates of multiple numbers.