Merging any 2 arrays and sorting the whole
Pretty self explanatory. We merge 2 arrays and then do a sorting algorithm on the merged array.

Note:
There is also a built in method in python that allows us to merge two arrays without even appending each element via a for loop. It is called .extend()

This is the advantage of lists and array lists ( on java ) over normal arrays as we needed to have created another separate array in that case to append all the elements of the first as well as the second array to the new array. 

We could also directly append the element in a sorted fashion (which would not be optimal here) which is explored in the "swalpaJargtira.py" code.