Insert an element in a specific position
Imagine you are in a marriage function and have sat down to have food, but your friend is somewhere in the middle of the line. So you ask the people sitting to more a step to the right to make space for you to sit with your friend and eat. This is the problem. 

Takeaways:
We take note of where the person wants to sit and then we traverse the array in reverse until that position and while we do, we go on appending the array to its next position i.e., a[i+1] = a[i]. When we finally gget to the position the person wants to sit in, we insert that person in the now free position.

Note: 
This can only be done while traversing the array on reverse. If we traverse normally, we would append the current iteration element with its +1, but then we would actually GO to that +1 iteration in the next iteration and that element would just be copied throughout the whole array. This is not what we want now do we. Still, we can use that as a means to copy a certain element throughout the array though.