Creates a 2D Jagged list from user input
This one was a bit tricky as it was different from Java and C++ quite a lot. This is due to there not being traditional arrays in Python but there only being lists. Lists are excellent but take some time getting used to when you come from Java or C++.  💢 

Key Takeaways:
1. This is how it works: You create single-individual arrays and append it to the main array where you want all arrays to be stored in. Essentially making it a 2D array. (List of lists).
2. IMPORTANT : Note that you need to re-initialize the variable you're using to store the arrays which you're gonna append to the main array because if not done so, after each loop the new entries (for the new array which you wanna append) will actually be appended to the previous jagged array u appended to the main array. So for example, if you had appended [1,2,3] and now wanna append [7,8] in the next jagged loop, what actually happens is the first jagged will be [1,2,3] and the second jagged will be [1,2,3,7,8]. So yep be careful about that. Print statement given for more clarity on this in code.

Further:
This can also be made into any dimensional array as needed by nesting the loops further, appending to the jagged level earlier and then appending all those lists of lists to the main list (3 dimensional list)

😄