Gives the biggest sub-array (sub-list)
-------------------- Kadane's algorithm ----------------
Kadane knows how to find the biggest subarray in any given array cuz he has an algorithm that he uses. 

Me : So Mr.Kadane, can you please tell us how you do it 🤔❓ 
Kadane : Sure! Here's how! 😄 

Key Takaways:
Here, you gotta do a traversal and go on adding each element to a variable and check if that variable constantly stays above zero. Why? Because if that reaches zero, that means all the numbers before it add up to nothing or less. So its a waste to get them to the biggest subarray list. So we reset (only) the sum variable and the array to which we appended the maximum list elements and continue traversing the list. At the end we will be left with a sum that would be the maximum sum and a subarray (or rather, sublist) containing the max subarray contained within the original array.

Kadane : Got it? 😄 
Me : Noice! I got it! 😃👍 

🥳 🥳 🎆