Finds the pivot element (explained in desc)
A pivot element is such an element in an array that all the elements before it should add up to all the elements after it. Easy, but deceptively so.

Key Takeaways:
Tried to initially do it by how we do it on pen and paper. By running a loop forward, adding each element and running another j loop which would run in reverse, adding elements UNTIL that i element. The problem to this approach is that the elements would JUST miss each other. That is, the pivot numbers we would get on different index but next to each other. For example, lets say index 3 on the ith loop, and index 2 on the jth loop. So it was a nit of a headache as can be imagined. 

The next approach was the one I've used here as well as many others have as well. That is, get the total sum by running the reverse loop, then run a forward loop and compare by adding to the forward sum 'forsum' and removing from the reverse sum 'revsum'. The thing here is to have a good understanding of loops because we need to compare after reduction from revsum, THEN add to forsum if match not found. Otherwise you'd  miss it by an index yet again.