Prints any string N times
This prints any string you give it as many times you want it to using recursion.

Key takeaways:
Here in range, 
the first argument is the starting position (which you can give anything basically as long as its lesser than the base condition), 
the second argument is the base condition (how many times you want it to print before returning)
the third argument is the string which you'd print of-course.

Yay! basic recursion!