Get the sum till n using 2 parameters only
Gets the sum without using 3 parameters like before. (uses only 2 parameters, one to count down till zero to iterate the recursion, other to hold the sum in the res variable)

Note: The tricky part here that needs to be realized is that line 11 is necessary. Without it, if you instead try to pass rec(i-=1,res+i), the i the res variable will take will be the already -1'd value of i because if i-=1. 
This can also be avoided by changing the function to be like rec(res+1, i-=1) but this was a learning point, so I have included it here. 😸