Factorial(n), but functional (return value based)
Gives the factorial of a number using return value of recursions. i.e., line 9 waits for rec(i-1) to return a value. this causes recursion to happen until it goes to 0 where, it is returned a 1. While backtracking, that 1 multiplies with 2 that whole answer multiplies with 3 and so on until n. Until finally the original i+rec(i-1) is reached on line 9 where we will be getting the answer for the factorial. This will be returned to the mail function where it will be printed.